backend/requirements.txt
aerich
asyncpg
fastapi
passlib
python-jose
python-multipart
tortoise-orm
uvicorn[standard]
